Fix a nasty dag combiner bug that caused nondeterminstic crashes (MY FAVORITE!): SimplifySelectOps would eliminate a Select, delete it, then return true.
The clients would see that it did something and return null.

The top level would see a null return, and decide that nothing happened,
proceeding to process the node in other ways: boom.

The fix is simple: clients of SimplifySelectOps should return the select
node itself.

In order to catch really obnoxious boogs like this in the future, add an
assert that nodes are not deleted.  We do this by checking for a sentry node
type that the SDNode dtor sets when a node is destroyed.

llvm-svn=28514