Added a work in the DWARF parser when we parse an array that ends up having no elements so that they at least have 1 element.
Added the ability to show the declaration location of variables to the 
"frame variables" with the "--show-declaration" option ("-c" for short).

Changed the "frame variables" command over to use the value object code
so that we use the same code path as the public API does when accessing and
displaying variable values.

llvm-svn=113733