Initial rough-in for the LandingPadInstr.
The OperandList is a list of the Values it can catch. There is a private vector
that has an enum that's used to decrypt the Value. It's a bit inelegant, but was
done this way so that the clauses could "interleave." E.g., you could have a
catch-filter-catch sequence.

It doesn't have any functionality just yet.

llvm-svn=135984