Fix detection of whether or not symbols are from j3 or not.
Unfortunately, native code that's part of OpenJDK likes to dlopen things with
RTLD_GLOBAL, resulting in the existing mechanism for determining where a symbol
came from inaccurate.

Instead, search both our process (and any global libraries), but also search our
loaded libraries.

Using this, we can determine which symbols are in vmkit: they must be in in the self
search AND be a different symbol from what was found in the library search.

llvm-svn=144857