elf-core: Remove SetTriple workaround
On FreeBSD I see that Arch, Vendor, OS, and Environment are unchanged
after this call (for x86_64, at least), and core debugging works without
it.

If we need to restore it we should be able to switch on
arch.GetTriple.getOS() instead of a compile-time #ifdef.

Thanks to Daniel Malea for testing on Linux.

llvm-svn=186862