Fix c/InternVariableWeakref.c.
The test was failing when building with gcc 4.8. It looks like weakrefs are
reachable via emit_varpool_aliases and emit_varpool_weakref. This causes us to
try to use rauw twice, causing an assert when we get a this->rauw(this).

The patch simply skips weakrefs in emit_varpool_aliases.

llvm-svn=208010