[llgo] irgen: always use TargetMachine's data layout
Summary:
Another attempt at resolving the runtime assertion
in llgoi due to data layout mismatch between module
and execution engine.

The X86 data layout constant appears to be unnecessary,
and does not match what the execution engine picks.
Using the registered Target, we pick the same data
layout as the execution engine.

While I was in the vicinity, I deleted the last
remnants of PNaCl support.

Reviewers: pcc

Subscribers: jfb, llvm-commits, dschuff

Differential Revision: http://reviews.llvm.org/D12630

llvm-svn=248565