[MSSA] Fix PR28632 in the 3.9 branch.
The now-removed assertion was really more for initial debugging; it's
perfectly valid (albeit relatively rare) for `Q.Visited.size()` to be
greater than 1,000. A similar patch hasn't been applied to trunk
because the piece of code this assertion was a part of no longer exists
in trunk.

Fix okayed by Danny, merge to 3.9 okayed by Hans.

llvm-svn=276337